rocketdata/requirements.txt
Django>=3.2.4
django-filter>=2.4.0
djangorestframework>=3.12.4
djangorestframework-api-key>=2.0.0
markdown>=3.3.4
psycopg2>=2.8.6